Completes maze in 1381 moves
